added menu
It ain't no joke, I'd like to buy the world a toke
And teach the world to sing in perfect harmony
And teach the world to snuff the fires and the liars
Hey, I know it's just a song but it's spice for the recipe
This is a love attack, I know, went out but it's back
It's just like any fad, it retracts before impact
And just like fashion, it's a passion for the with it and hip
If you got the goods, they'll come and buy it just to stay in the clique
So don't delay, act now, supplies are running out
Allow if you're still alive, six to eight years to arrive
And if you follow, there may be a tomorrow
But if the offer is shun, you might as well be walkin' on the sun
Twenty-five years ago, they spoke out and they broke out
Of recession and oppression and together they toked
And they folked out with guitars around a bonfire
Just singin' and clappin', man, what the hell happened there?
Some were spellbound, some were hell bound
Some they fell down and some got back up
And fought back against the melt-down
And their kids were hippie chicks or hypocrites
Because fashion is smashin' the true meaning of it
So don't delay, act now, supplies are running out
Allow if you're still alive, six to eight years to arrive
And if you follow, there may be a tomorrow
But if the offer is shun, you might as well be walkin' on the sun
And it ain't no joke when our mama's handkerchief is soaked
With her tears because her baby's life has been revoked
The bond is broke up, so choke up and focus on the close up
Mr. Wizard can't reform, no God-like hocus-pocus
So don't sit back, kick back and watch the world get bushwhacked
News at ten, your neighborhood is under attack
Put away the crack before the crack put you away
You need to be there when your baby's old enough to relate
So don't delay, act now, supplies are running out
Allow if you're still alive, six to eight years to arrive
And if you follow, there may be a tomorrow
But if the offer is shun, you might as well be walkin' on the sun